Switch for the job_util option
